requirements.txt
asgiref==3.4.1
async-timeout==4.0.2
cassandra-driver==3.29.2
certifi==2024.8.30
charset-normalizer==2.0.12
click==8.0.4
dataclasses==0.8
Django==3.2.25
django-cassandra-engine==1.6.3
django-redis==5.4.0
django-user-agents==0.4.0
geomet==0.2.1.post1
gevent==22.10.2
greenlet==2.0.2
gunicorn==21.2.0
h11==0.13.0
idna==3.10
importlib-metadata==4.8.3
packaging==21.3
Pillow==8.4.0
psycopg2-binary==2.9.5
pyparsing==3.1.4
pytz==2024.2
redis==4.3.6
requests==2.27.1
six==1.16.0
sqlparse==0.4.4
stripe==11.1.0
typing_extensions==4.1.1
ua-parser==0.18.0
urllib3==1.26.20
user-agents==2.2.0
uvicorn==0.16.0
zipp==3.6.0
zope.event==4.6
zope.interface==5.5.2
